+++ yash-semantics/src/runner.rs
@@ -79,20 +79,17 @@ use yash_syntax::syntax::List;
 /// ```
 /// # futures_executor::block_on(async {
 /// # use std::cell::RefCell;
-/// # use std::num::NonZeroU64;
 /// # use std::ops::ControlFlow::Continue;
-/// # use std::rc::Rc;
 /// # use yash_env::Env;
 /// # use yash_env::input::Echo;
 /// # use yash_semantics::ExitStatus;
 /// # use yash_semantics::read_eval_loop;
 /// # use yash_syntax::input::Memory;
 /// # use yash_syntax::parser::lex::Lexer;
-/// # use yash_syntax::source::Source;
 /// let mut env = Env::new_virtual();
 /// let mut ref_env = RefCell::new(&mut env);
 /// let input = Box::new(Echo::new(Memory::new("case foo in (bar) ;; esac"), &ref_env));
-/// let mut lexer = Lexer::new(input, NonZeroU64::MIN, Rc::new(Source::CommandString));
+/// let mut lexer = Lexer::new(input);
 /// let result = read_eval_loop(&ref_env, &mut lexer).await;
 /// drop(lexer);
 /// assert_eq!(result, Continue(()));
@@ -206,7 +203,6 @@ mod tests {
     use crate::tests::echo_builtin;
     use crate::tests::return_builtin;
     use futures_util::FutureExt;
-    use std::num::NonZeroU64;
     use std::rc::Rc;
     use yash_env::input::Echo;
     use yash_env::input::Memory;
@@ -294,8 +290,7 @@ mod tests {
         env.options.set(Verbose, On);
         let ref_env = RefCell::new(&mut env);
         let input = Box::new(Echo::new(Memory::new("case _ in esac"), &ref_env));
-        let line = NonZeroU64::new(1).unwrap();
-        let mut lexer = Lexer::new(input, line, Rc::new(Source::CommandString));
+        let mut lexer = Lexer::new(input);
 
         let result = read_eval_loop(&ref_env, &mut lexer).now_or_never().unwrap();
         drop(lexer);
@@ -410,9 +405,7 @@ mod tests {
             }
         }
 
-        let input = Box::new(BrokenInput);
-        let line = NonZeroU64::new(1).unwrap();
-        let mut lexer = Lexer::new(input, line, Rc::new(Source::Unknown));
+        let mut lexer = Lexer::new(Box::new(BrokenInput));
         let mut env = Env::new_virtual();
         let ref_env = RefCell::new(&mut env);